Add files via upload
This lab practices logic design using familiar building blocks. The goal of this lab is for you to implement a circuit that generates square waves at specific frequencies based on inputs, and outputs them to headphones using the audio jack present on the Real Digital Blackboard. When you successfully complete this lab, you will have developed a piece of intellectual property that you might be able to re-use in the future.